+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Added
+- `RunnerSettings.filePathPrintBase` to control how file paths are printed.
+  - Required when C# scripts (.csx) are used.
+  - https://github.com/StateSmith/StateSmith/issues/79
+
+
 ## [0.7.3-alpha]
 ### Added
 - draw.io - add support for draw.io plugin and new drawing style where a state can have its text event handlers as a child text node instead of its label #77.